backend/requirements.txt
fastapi
uvicorn
python-multipart
onnxruntime
numpy
chromadb
requests
python-dotenv
huggingface_hub
pypdf
pygithub
transformers
notion-client
